Tolerance the deviation in headroom watermark
In the headroom watermark test, it is expected that the watermark is multiple of cell size.
However, there is some deviation caused by the below reasons:
- The headroom size in SONiC is in bytes and isn't the multiple of cell size.
- The SAI calculate the headroom watermark as "sdk max occupancy" - "sonic configuration",
  if the value configured in SONiC doesn't aligned with cell size boundary, the watermark
  isn't cell size aligned.
Especially, when the packets start to trek to headroom (xoff zone), the headroom size isn't 0
(but it should be), which fails the headroom watermark test.
To tolerance the deviation, the number of packets to occupy all the shared buffer
is a bit less than it should be.

Signed-off-by: Stephen Sun <stephens@nvidia.com>