[qos] Enlarge the margin for the PFCXontest on Mellanox ACS-MSN4700 400G port (#2611)
* Enlarge the margin of the PFCXontest on 4700 400G port

What is the motivation for this PR?
Originally we had a margin but in some cases, the test can still fail at a very low probability
So we need to enlarge it to make it safer.
The common infrastructure of PFCXontest has been adjusted a bit regarding the usage of margin as well. As the margin is calculated on a per-platform basis, for other platforms without 400G port won't be affected.

Change-Id: I4555dedd854e2802f5f198f2f10b4710c2990343
Signed-off-by: Stephen Sun stephens@nvidia.com