Fix add-topo error RTNETLINK answers: No route to host (#2745)
Summary: Fix add-topo error RTNETLINK answers: No route to host.
This issue occurs if a ptf-docker does not have a ipv6 address, but the server file where the ptf-docker is to be started has an ipv6 mgmt gw address defined.
Modified the check for adding ipv6 default gw.
Verified add-topo on KVM and impacted testbeds