Bug fix - handle the cleanup only case without image URL (#2783)
Description of PR
Summary: Fix bug introduced by #2661 where if new_image_url is not provided, the installation logic still proceeds.
This leads to error in the Ansible module in upgrade_sonic.yml step for removing old sonic images.

Approach
What is the motivation for this PR?
How did you do it?
Added the check for no image URL. And do not execute installation logic if URL is None.

How did you verify/test it?
Tested on a physical lab device, and the Ansible execution of upgrade_sonic.yml passed.