[pretest] Add support to update python saithrift package on ptf (#2793)
Description of PR
Some tests like qos_sai are dependent on the python saithrift package installed on the ptf. This PR adds support to install the correct python saithrift package based on the image version being tested prior to the test run

How did you do it?
Added --py_saithrift_url option in conftest.py
Changes in test_pretest to install the package if the url is present and valid. In case of missing url or download error, the testcase is skipped

How did you verify/test it?
Tested without specifying --py_saithrift_url and testcase was skipped
Tested with incorrect url and testcase was skipped
Tested with valid url and package was installed on ptf

Signed-off-by: Neetha John <nejo@microsoft.com>