Ignore None to int error for both python2 and python3 (#2830)
When cast None type to int type, python3 and python2 show different error. In python2, it shows:
int() argument must be a string or a number, not 'NoneType'
While in python3 it shows:
int() argument must be a string, a bytes-like object or a number, not 'NoneType'
As thermalctd in master branch has been migrated to python3, and 201911 branch is still using python2, we need handle both cases when ignoring such error in log analyzer.