Fix BUG #2601416: QoS test failed due to "IndexError: list index out of range"
Currently, the dynamic buffer calculation is enabled by a script after the DUT is deployed.
One step to enable it is to transform all lossless PGs.
However, sometimes the lossless PGs haven't be created when the script is running, which causes no lossless PG configured in the DUT, which in turn fails the test

Fix: if there is no lossless PG configured, lossless PGs will be created for all ports

Change-Id: I8b74c8ffab272015965100f713033e1048f45b1d
Signed-off-by: Stephen Sun <stephens@nvidia.com>